requirements.txt
cycler==0.10.0
guizero==1.2.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
opencv-python==4.5.4.58
pandas==1.3.4
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
